# docs/docSwagger.yaml
openapi: 3.0.3
info:
  title: AParkApp API
  version: 1.0.0
  description: Documentación OpenAPI de los endpoints disponibles en la carpeta feature.
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    VehicleTypeEnum:
      type: string
      enum: [Car, MotorCycle, Truck, SUV, Custom]
...
    DriverResponse:
      type: object
      additionalProperties: false
      required: [id, name, dni, phone]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        dni:
          type: string
        phone:
          type: string
      example:
        id: 550e8400-e29b-41d4-a716-446655440001
        name: Carlos Gomez
        dni: 12345678
        phone: 3001234567
    VehicleDTO:
      type: object
      additionalProperties: false
      required: [id, plate, model, color, brand, userId, vehicleType]
      properties:
        id:
          type: string
          format: uuid
        plate:
          type: string
        model:
          type: string
        color:
          type: string
        note:
          type: string
        brand:
          type: string
        userId:
          type: string
          format: uuid
        vehicleType:
          $ref: '#/components/schemas/VehicleTypeEnum'
      example:
        id: 550e8400-e29b-41d4-a716-446655440010
        plate: ABC123
        model: Corolla
        color: White
        note: Electric windows
        brand: Toyota
        userId: 550e8400-e29b-41d4-a716-446655440000
        vehicleType: Car
    NewVehicleDTO:
      type: object
      additionalProperties: false
      required: [plate, userId, model, color, brand, vehicleType]
      properties:
        plate:
          type: string
        userId:
          type: string
          format: uuid
        model:
          type: string
        color:
          type: string
        note:
          type: string
        brand:
          type: string
        vehicleType:
          $ref: '#/components/schemas/VehicleTypeEnum'
      example:
        plate: ABC123
        userId: 550e8400-e29b-41d4-a716-446655440000
        model: Corolla
        color: White
        note: Electric windows
        brand: Toyota
        vehicleType: Car
    VehicleTypeDTO:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
      example:
        name: Car
    ParkingLotRequest:
      type: object
      additionalProperties: false
      required: [name, addressId, capacity, ownerId]
      properties:
        name:
          type: string
        addressId:
          type: string
          format: uuid
        capacity:
          type: integer
          format: int32
          minimum: 1
        ownerId:
          type: string
          format: uuid
      example:
        name: Centro
        addressId: 550e8400-e29b-41d4-a716-446655440100
        capacity: 120
        ownerId: 550e8400-e29b-41d4-a716-446655440000
    ParkingLotResponse:
      type: object
      additionalProperties: false
      required: [id, name, ownerId, ownerUsername, addressId, capacity, active]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ownerId:
          type: string
          format: uuid
        ownerUsername:
          type: string
        addressId:
          type: string
          format: uuid
        capacity:
          type: integer
          format: int32
        active:
          type: boolean
      example:
        id: 550e8400-e29b-41d4-a716-446655440200
        name: Centro
        ownerId: 550e8400-e29b-41d4-a716-446655440000
        ownerUsername: admin@example.com
        addressId: 550e8400-e29b-41d4-a716-446655440100
        capacity: 120
        active: true
    ParkingLotClientView:
      type: object
      additionalProperties: false
      required: [name, street, number, zipCode, capacity]
      properties:
        name:
          type: string
        street:
          type: string
        number:
          type: integer
          format: int32
        zipCode:
          type: string
        capacity:
          type: integer
          format: int32
      example:
        name: Centro
        street: Main Street
        number: 123
        zipCode: 110111
        capacity: 120
...
    TicketResponseDTO:
      type: object
      additionalProperties: false
      required: [id, reservationId, amount, paid, timeStamp, note]
      properties:
        id:
          type: string
          format: uuid
        reservationId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        paid:
          type: number
          format: double
        timeStamp:
          type: string
          format: date-time
        note:
          type: string
      example:
        id: 550e8400-e29b-41d4-a716-446655440700
        reservationId: 550e8400-e29b-41d4-a716-446655440500
        amount: 15000.0
        paid: 15000.0
        timeStamp: 2026-06-16T12:00:00
        note: Pago confirmado
    PaidRequestDTO:
      type: object
      additionalProperties: false
      required: [id, pay]
      properties:
        id:
          type: string
          format: uuid
        pay:
          type: number
          format: double
      example:
        id: 550e8400-e29b-41d4-a716-446655440700
        pay: 15000.0
paths:
...
  /drivers/{dni}/vehicles/{plate}:
    post:
      tags: [Conductores]
      security:
        - bearerAuth: []
      summary: Asocia un vehículo a un conductor
      parameters:
        - name: dni
          in: path
          required: true
          schema:
            type: string
        - name: plate
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Asociación realizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverResponse'
    delete:
      tags: [Conductores]
      security:
        - bearerAuth: []
      summary: Desasocia un vehículo de un conductor
      parameters:
        - name: dni
          in: path
          required: true
          schema:
            type: string
        - name: plate
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Asociación eliminada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverResponse'
...
  /vehicles/{vehicleId}:
    get:
      tags: [Vehículos]
      security:
        - bearerAuth: []
      summary: Obtiene un vehículo por id
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vehículo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDTO'
    put:
      tags: [Vehículos]
      security:
        - bearerAuth: []
      summary: Actualiza un vehículo
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVehicleDTO'
      responses:
        '200':
          description: Vehículo actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDTO'
    delete:
      tags: [Vehículos]
      security:
        - bearerAuth: []
      summary: Elimina un vehículo
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Vehículo eliminado
  /vehicle-type:
    get:
      tags: [Tipos de vehículo]
      security:
        - bearerAuth: []
      summary: Lista los tipos de vehículo
      responses:
        '200':
          description: Lista de tipos de vehículo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleTypeDTO'
    post:
      tags: [Tipos de vehículo]
      security:
        - bearerAuth: []
      summary: Crea un tipo de vehículo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleTypeDTO'
      responses:
        '201':
          description: Tipo de vehículo creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleTypeDTO'
  /vehicle-type/{vehicleTypeId}:
    get:
      tags: [Tipos de vehículo]
      security:
        - bearerAuth: []
      summary: Obtiene un tipo de vehículo por id
      parameters:
        - name: vehicleTypeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tipo de vehículo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleTypeDTO'
    put:
      tags: [Tipos de vehículo]
      security:
        - bearerAuth: []
      summary: Actualiza un tipo de vehículo
      parameters:
        - name: vehicleTypeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleTypeDTO'
      responses:
        '200':
          description: Tipo de vehículo actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleTypeDTO'
    delete:
      tags: [Tipos de vehículo]
      security:
        - bearerAuth: []
      summary: Elimina un tipo de vehículo
      parameters:
        - name: vehicleTypeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Tipo de vehículo eliminado
  /ParkingLot:
    get:
      tags: [Parqueaderos]
      security:
        - bearerAuth: []
      summary: Lista todos los parqueaderos
      responses:
        '200':
          description: Lista de parqueaderos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingLotResponse'
    post:
      tags: [Parqueaderos]
      security:
        - bearerAuth: []
      summary: Crea un parqueadero
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingLotRequest'
      responses:
        '201':
          description: Parqueadero creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingLotResponse'
  /ParkingLot/active:
    get:
      tags: [Parqueaderos]
      security:
        - bearerAuth: []
      summary: Lista los parqueaderos activos
      responses:
        '200':
          description: Parqueaderos activos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingLotResponse'
  /ParkingLot/{id}:
    put:
      tags: [Parqueaderos]
      security:
        - bearerAuth: []
      summary: Actualiza un parqueadero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingLotRequest'
      responses:
        '200':
          description: Parqueadero actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingLotResponse'
    delete:
      tags: [Parqueaderos]
      security:
        - bearerAuth: []
      summary: Elimina un parqueadero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Parqueadero eliminado
  /ParkingLot/{id}/restore:
    patch:
      tags: [Parqueaderos]
      security:
        - bearerAuth: []
      summary: Restaura un parqueadero eliminado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Parqueadero restaurado
  /ParkingLot/client:
    get:
      tags: [Parqueaderos]
      security:
        - bearerAuth: []
      summary: Lista los parqueaderos visibles para cliente
      responses:
        '200':
          description: Vista cliente de parqueaderos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingLotClientView'
...
  /tickets:
    get:
      tags: [Tickets]
      security:
        - bearerAuth: []
      summary: Lista todos los tickets
      responses:
        '200':
          description: Lista de tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketResponseDTO'
  /tickets/today:
    get:
      tags: [Tickets]
      security:
        - bearerAuth: []
      summary: Lista los tickets de hoy
      responses:
        '200':
          description: Lista de tickets del día
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketResponseDTO'
  /tickets/pay:
    post:
      tags: [Tickets]
      security:
        - bearerAuth: []
      summary: Registra el pago de un ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaidRequestDTO'
      responses:
        '200':
          description: Ticket pagado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponseDTO'
  /tickets/{id}:
    delete:
      tags: [Tickets]
      security:
        - bearerAuth: []
      summary: Cancela un ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Ticket cancelado