# src/main/java/com/equipo7/AParkApp/feature/ticket/TicketController.java
package com.equipo7.AParkApp.feature.ticket;

import com.equipo7.AParkApp.feature.ticket.domain.dto.PaidRequestDTO;
import com.equipo7.AParkApp.feature.ticket.domain.dto.TicketResponseDTO;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.UUID;

@RestController
@RequestMapping("/tickets")
@RequiredArgsConstructor
@Tag(name = "Tickets", description = "Gestión de tickets y pagos")
@SecurityRequirement(name = "bearerAuth")
public class TicketController {
    private final TicketService service;

    @PreAuthorize("hasAnyRole('OWNER', 'EMPLOYEE')")
    @GetMapping
    @ResponseStatus(HttpStatus.OK)
    @Operation(summary = "Lista todos los tickets")
    public List<TicketResponseDTO> findAll() {
        return service.findAll();
    }

    @PreAuthorize("hasAnyRole('OWNER', 'EMPLOYEE')")
    @GetMapping("/today")
    @ResponseStatus(HttpStatus.OK)
    @Operation(summary = "Lista los tickets de hoy")
    public List<TicketResponseDTO> findToday() {
        return service.findToday();
    }

    @PreAuthorize("hasAnyRole('OWNER', 'EMPLOYEE','CLIENT')")
    @PostMapping("/pay")
    @ResponseStatus(HttpStatus.OK)
    @Operation(summary = "Registra el pago de un ticket")
    public TicketResponseDTO addPay(@Valid @RequestBody PaidRequestDTO request) {
        return service.addPay(request);
    }

    @PreAuthorize("hasAnyRole('OWNER', 'EMPLOYEE')")
    @DeleteMapping("/{id}")
    @ResponseStatus(HttpStatus.OK)
    @Operation(summary = "Cancela un ticket")
    public void cancel(@PathVariable UUID id) {
        service.delete(id);
    }
}